+++ app/Validators/Request/Fields/ItemCategory.php
@@ -1,22 +1,22 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Http\Parameters\Request\Validators;
+namespace App\Validators\Request\Fields;
 
-use App\Http\Parameters\Request\Validators\Validator as BaseValidator;
+use App\Validators\Request\Fields\Validator as BaseValidator;
 use Illuminate\Contracts\Validation\Validator;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Validator as ValidatorFacade;
 
 /**
- * Validation helper class for resource types, returns the generated validator objects
+ * Validation helper class for item category, returns the generated validator objects
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2019
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ResourceType extends BaseValidator
+class ItemCategory extends BaseValidator
 {
     /**
      * Return the validator object for the create request
@@ -27,15 +27,16 @@ class ResourceType extends BaseValidator
      */
     public function create(Request $request): Validator
     {
-        $messages = [];
-        foreach (Config::get('api.resource-type.validation.POST.messages') as $key => $custom_message) {
-            $messages[$key] = trans($custom_message);
-        };
+        $decode = $this->hash->category()->decode($request->input('category_id'));
+        $category_id = null;
+        if (count($decode) === 1) {
+            $category_id = $decode[0];
+        }
 
         return ValidatorFacade::make(
-            $request->all(),
-            Config::get('api.resource-type.validation.POST.fields'),
-            $this->translateMessages('api.resource-type.validation.POST.messages')
+            ['category_id' => $category_id],
+            Config::get('api.item-category.validation.POST.fields'),
+            $this->translateMessages('api.item-category.validation.POST.messages')
         );
     }
 }
