+++ app/Validators/Request/Fields/ResourceType.php
@@ -1,22 +1,22 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Http\Parameters\Request\Validators;
+namespace App\Validators\Request\Fields;
 
-use App\Http\Parameters\Request\Validators\Validator as BaseValidator;
+use App\Validators\Request\Fields\Validator as BaseValidator;
 use Illuminate\Contracts\Validation\Validator;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Validator as ValidatorFacade;
 
 /**
- * Validation helper class for item category, returns the generated validator objects
+ * Validation helper class for resource types, returns the generated validator objects
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2019
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ItemCategory extends BaseValidator
+class ResourceType extends BaseValidator
 {
     /**
      * Return the validator object for the create request
@@ -27,16 +27,15 @@ class ItemCategory extends BaseValidator
      */
     public function create(Request $request): Validator
     {
-        $decode = $this->hash->category()->decode($request->input('category_id'));
-        $category_id = null;
-        if (count($decode) === 1) {
-            $category_id = $decode[0];
-        }
+        $messages = [];
+        foreach (Config::get('api.resource-type.validation.POST.messages') as $key => $custom_message) {
+            $messages[$key] = trans($custom_message);
+        };
 
         return ValidatorFacade::make(
-            ['category_id' => $category_id],
-            Config::get('api.item-category.validation.POST.fields'),
-            $this->translateMessages('api.item-category.validation.POST.messages')
+            $request->all(),
+            Config::get('api.resource-type.validation.POST.fields'),
+            $this->translateMessages('api.resource-type.validation.POST.messages')
         );
     }
 }
