+++ app/Validators/Request/Fields/SubCategory.php
@@ -1,41 +1,41 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Http\Parameters\Request\Validators;
+namespace App\Validators\Request\Fields;
 
-use App\Http\Parameters\Request\Validators\Validator as BaseValidator;
+use App\Validators\Request\Fields\Validator as BaseValidator;
 use Illuminate\Contracts\Validation\Validator;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Validator as ValidatorFacade;
 
 /**
- * Validation helper class for resources, returns the generated validator objects
+ * Validation helper class for sub categories, returns the generated validator objects
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2019
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class Resource extends BaseValidator
+class SubCategory extends BaseValidator
 {
     /**
-     * Create the validation rules for the create request
+     * Create the validation rules for the create (POST) request
      *
-     * @param integer $resource_type_id
+     * @param integer $category_id
      *
      * @return array
      */
-    private function createRules(int $resource_type_id): array
+    private function createRules(int $category_id): array
     {
         return array_merge(
             [
                 'name' => [
                     'required',
                     'string',
-                    'unique:resource,name,null,id,resource_type_id,' . $resource_type_id
+                    'unique:sub_category,name,null,id,category_id,' . $category_id
                 ],
             ],
-            Config::get('api.resource.validation.POST.fields')
+            Config::get('api.subcategory.validation.POST.fields')
         );
     }
 
@@ -43,16 +43,16 @@ class Resource extends BaseValidator
      * Return the validator object for the create request
      *
      * @param Request $request
-     * @param integer $resource_type_id
+     * @param integer category_id
      *
      * @return Validator
      */
-    public function create(Request $request, int $resource_type_id): Validator
+    public function create(Request $request, int $category_id): Validator
     {
         return ValidatorFacade::make(
             $request->all(),
-            self::createRules($resource_type_id),
-            $this->translateMessages('api.resource.validation.POST.messages')
+            self::createRules($category_id),
+            $this->translateMessages('api.subcategory.validation.POST.messages')
         );
     }
 }
